Privileges fixed with nice neat indentation
Adjusted the 'user' privileges to only grant and not revoke privleges.
This makes it idempotent where before if the same username was used for
an admin and the user, the last action was to revoke the 'super'
privilege which left replication access broken.
(cherry picked from commit dde714e2fab218bf9dee526cc528736e82d4127b)