Appropriate privileges added only
The previous code when used in our templates caused replication to be
broken when using the same username and password for admin /
replication and the application DB users.

The issue was caused by the use of revoke which instead of setting the
appropriate privileges removed the super privilege from the user
removing the ability to set the master host. i.e not idempotent. This
is resolved by just granting the appropriate ones and therefore not
revoking required ones.